# optimize db: add index, rewrite query